backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
python-docx-template==0.17.0
sqlalchemy==2.0.25
flask-sqlalchemy==3.1.1
marshmallow==3.20.2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9